refactor: Add program parameter to the prereqs of courses
To handle cases where a course is shared with multiple programs but for each program there are different prereq courses.
ELCN203 for example which is shared with cce, eee, mee and each program has its own prereq for this course.
However they all share a prereq course which is MTHN102, so the parameter would be 'ALL' for this case to handle general cases.